Shift tests out of deprecated
